feat: plan list 분리
