+++ src/router/index.js
@@ -4,6 +4,7 @@ import HomeView from "../views/home/HomeView.vue";
 import LoginView from "@/views/user/LoginView.vue";
 import PlanView from "@/views/plan/PlanView.vue";
 import PlanDetailView from "@/views/plan_detail/PlanDetailView.vue";
+import PlanListView from "@/components/plan/PlanListView.vue";
 import { useAuthStore } from "@/stores/auth";
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -34,6 +35,13 @@ const router = createRouter({
       path: "/plan",
       name: "Plan",
       component: PlanView,
+      children: [
+        {
+          path: "list",
+          name: "PlanList",
+          component: PlanListView,
+        },
+      ],
     },
     {
       path: "/plan/:id",
